Profile Update, email and password change
Summary:
Added email and password changing as well as updated the profile view to be allow better interaction.

Will be adding this kind of UI to all other components in the app
Fixes T67

Test Plan:
I have no means of testing if the email change actually works or not since I am not on the prod environment as far as the server goes. Meaning it doesn't send an email because it doesn't know how to send it

Test password editing

Reviewers: Info-Screen

Reviewed By: Info-Screen

Subscribers: WyldBot, Info-Screen

Maniphest Tasks: T67

Differential Revision: https://phabricator.wildfyre.net/D58